parser halfway done

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: bamrouch <bamrouch@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 15:42:21 by bamrouch          #+#    #+#             */
-/*   Updated: 2022/11/18 17:55:40 by bamrouch         ###   ########.fr       */
+/*   Updated: 2023/02/01 13:35:27 by bamrouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ static int	ft_isspace(int c)
 	return (0);
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int				i;
 	unsigned long	res;
